# Use css style on component testing