# Description

Some component tests rely on css class in order to determine the component properties. For example in https://github.com/trento-project/web/commit/56a978b615e4cdd191f5ddd37500d57ff262d293 the `toBeVisible()` matcher   look for style `display: none` to be applied to the element or one of its parents.

This PR proposes the `<WithStyle>` test utility. `<WithStyle>` is a higher-order component that prepends the application css to the component to be rendered. It's meant to be used only on tests that require such ability (for keeping the test footprint small).

In order to work the application css must be built already, thus a `pretest` script is added to package.json. 
~Furthermore, the package.json has been refactored a bit to make use of npm variables (not required, it can be reverted).~