@balanza 2024-06-18
A quick thought on performance based on my laptop

```sh
# running command
time npm run test js/common/Tooltip/Tooltip.test.jsx

# without style
npm run test js/common/Tooltip/Tooltip.test.jsx  3.43s user 0.76s system 180% cpu 2.325 total

# with style, generated just in time (first run)
npm run test js/common/Tooltip/Tooltip.test.jsx  5.33s user 0.96s system 168% cpu 3.727 total

# with style, already generated
npm run test js/common/Tooltip/Tooltip.test.jsx  3.48s user 0.73s system 173% cpu 2.419 total
```