@balanza 2024-06-19
Addendum: with @jamie-suse we investigated a bit around the possibility to load CSS once for all JSDOM instances. It seems there is no this possibility: furthermore, it seems this feature is considered out-of-scope by JSDOM itself (see https://github.com/testing-library/jest-dom/issues/70).